+++ src/main/java/lk/easycarrentalpvt/spring/repo/DamageRepo.java
@@ -3,5 +3,5 @@ package lk.easycarrentalpvt.spring.repo;
 import lk.easycarrentalpvt.spring.entity.Damage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DamageRepo extends JpaRepository<Damage,Integer> {
+public interface DamageRepo extends JpaRepository<Damage,Long> {
 }
