+++ src/main/java/lk/easycarrentalpvt/spring/repo/MaintainsRepo.java
@@ -4,5 +4,5 @@ package lk.easycarrentalpvt.spring.repo;
 import lk.easycarrentalpvt.spring.entity.Maintains;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MaintainsRepo extends JpaRepository<Maintains,Integer> {
+public interface MaintainsRepo extends JpaRepository<Maintains,Long> {
 }
