+++ src/main/java/lk/easycarrentalpvt/spring/repo/RentPaymentRepo.java
@@ -4,5 +4,5 @@ package lk.easycarrentalpvt.spring.repo;
 import lk.easycarrentalpvt.spring.entity.RentPayment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RentPaymentRepo extends JpaRepository<RentPayment,Integer> {
+public interface RentPaymentRepo extends JpaRepository<RentPayment,Long> {
 }
