+++ src/main/java/lk/easycarrentalpvt/spring/repo/RentReturnsRepo.java
@@ -4,5 +4,5 @@ package lk.easycarrentalpvt.spring.repo;
 import lk.easycarrentalpvt.spring.entity.RentReturns;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RentReturnsRepo extends JpaRepository<RentReturns,Integer> {
+public interface RentReturnsRepo extends JpaRepository<RentReturns,Long> {
 }
